delete Dist_col in calculation & fix descrition of GW_IDX
distance to collector was marked as "not used". In the original calculation, it was multiplied with 0.01 to minimize the effect.   Delete it from the formula. And will find how to include this measure in calculation in another way.
GW_IDX; 0 means no recharge area.   CON_GW, which = 1-GW_IDX, was used to calculate BASE_score. Higher BASE_score means more lightly to be developed.